app/requirements.txt
# Django's framework
Django>=4.0
psycopg2>=2.8

# Django Rest Framework
djangorestframework
markdown
django-filter

# Django Rest Auth
dj-rest-auth

# Test, Lyting and debugging
pytest-django
pylama
ptvsd
ipython

# Others
requests